cleaned up Program.cs
