+++ websmtp/Program.cs
@@ -1,8 +1,3 @@
-using System.Net;
-using System.Security.Cryptography.X509Certificates;
-using Microsoft.Data.Sqlite;
-using Microsoft.EntityFrameworkCore;
-using websmtp.Database;
 using websmtp.Startup;
 
 CommandLine.ParseStartupArgs(args);
@@ -22,25 +17,7 @@ CommandLine.ParseModifiersArgs(args, app);
 
 if (builder.Environment.IsEnvironment("Test"))
 {
-    using (var scope = app.Services.CreateScope())
-    {
-        var dbContext = scope.ServiceProvider.GetRequiredService<DataContext>();
-
-        if (dbContext.Database.GetDbConnection() is not SqliteConnection sqliteConnection)
-        {
-            throw new Exception("Test environement, but database does not appear to be Sqlite... Aborting");
-        }
-
-        var logger = scope.ServiceProvider.GetRequiredService<ILogger<Program>>();
-        logger.LogInformation("Test Environement, recreating TEST database.");
-        dbContext.Database.EnsureDeleted();
-        if (!dbContext.Database.EnsureCreated())
-        {
-            throw new Exception("Database not created");
-        }
-
-        var q = dbContext.RawMessages.ToList();
-    }
+    Startup.PrepareTestingEnvironement(app);
 }
 
 app.Run();
